Move DOM classes between funny folders
