+++ Source/OCompiler/Analyze/SemanticsV2/Dom/Expression/Call/BaseConstructorCallExpression.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using BaseToken = OCompiler.Analyze.Lexical.Tokens.Keywords.Base;
 
-namespace OCompiler.Analyze.SemanticsV2.Dom.Expression;
+namespace OCompiler.Analyze.SemanticsV2.Dom.Expression.Call;
 
 internal class BaseConstructorCallExpression : CallExpression
 {
