+++ Source/OCompiler/Analyze/SemanticsV2/Dom/Expression/Call/ObjectCreateExpression.cs
@@ -1,7 +1,6 @@
-using System.Collections;
 using System.Collections.Generic;
 
-namespace OCompiler.Analyze.SemanticsV2.Dom.Expression;
+namespace OCompiler.Analyze.SemanticsV2.Dom.Expression.Call;
 
 internal class ObjectCreateExpression : CallExpression
 {
