+++ Source/OCompiler/Analyze/SemanticsV2/Dom/Expression/ICanHaveParameters.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using OCompiler.Analyze.SemanticsV2.Dom.Expression.Special;
 
 namespace OCompiler.Analyze.SemanticsV2.Dom.Expression;
