+++ Source/OCompiler/Analyze/SemanticsV2/Dom/Expression/Special/ParameterDeclarationExpression.cs
@@ -1,14 +1,13 @@
-using System.Text;
 using OCompiler.Analyze.SemanticsV2.Dom.Type;
 using OCompiler.Analyze.SemanticsV2.Dom.Type.Member;
 
-namespace OCompiler.Analyze.SemanticsV2.Dom.Expression;
+namespace OCompiler.Analyze.SemanticsV2.Dom.Expression.Special;
 
 internal class ParameterDeclarationExpression : Expression
 {
     public TypeReference Type { get; set; }
     
-    public TypeMember? Holder { get; set; }
+    public new TypeMember? Holder { get; set; }
     
     public ParameterDeclarationExpression(string name, TypeReference type) : base(name)
     {
