+++ Source/OCompiler/Analyze/SemanticsV2/Tree/ParsedClassTree.Expressions.cs
@@ -4,7 +4,8 @@ using System.Linq;
 using OCompiler.Analyze.Lexical.Tokens;
 using OCompiler.Analyze.Lexical.Tokens.BooleanLiterals;
 using OCompiler.Analyze.Lexical.Tokens.Keywords;
-using OCompiler.Analyze.SemanticsV2.Dom.Expression;
+using OCompiler.Analyze.SemanticsV2.Dom.Expression.Call;
+using OCompiler.Analyze.SemanticsV2.Dom.Expression.NameReference;
 using OCompiler.Analyze.SemanticsV2.Dom.Expression.Primitive;
 using OCompiler.Analyze.Syntax.Declaration.Expression;
 using OCompiler.Exceptions;
